+++ agent/debug_test.go
@@ -5,7 +5,7 @@ package agent
 import (
 	"io/ioutil"
 
-	"github.com/Sirupsen/logrus"
+	"github.com/sirupsen/logrus"
 )
 
 func init() {