+++ backend/apid/graphql/relay/node.go
@@ -3,7 +3,7 @@ package relay
 import (
 	"context"
 
-	"github.com/Sirupsen/logrus"
+	"github.com/sirupsen/logrus"
 	"github.com/sensu/sensu-go/backend/apid/graphql/globalid"
 	"github.com/sensu/sensu-go/graphql"
 )