+++ backend/apid/routers/logger.go
@@ -1,6 +1,6 @@
 package routers
 
-import "github.com/Sirupsen/logrus"
+import "github.com/sirupsen/logrus"
 
 var logger = logrus.WithFields(logrus.Fields{
 	"component": "apid.routers",