+++ cli/client/client.go
@@ -6,7 +6,7 @@ import (
 	"net/http"
 	"time"
 
-	"github.com/Sirupsen/logrus"
+	"github.com/sirupsen/logrus"
 	"github.com/go-resty/resty"
 	"github.com/sensu/sensu-go/cli/client/config"
 )